Minor bug fix with file existence checks
Fixed the file existence checks for the pileup and hmmdata files, as
when more than one chromosome is being examined (e.g. not the toy
dataset), the results of the command substitutions in parent1or2-hmm.sh
contain whitespace characters which are then interpreted by the shell,
and exceed the unary nature of the test operator.

The quick fix applies to bash, and is to use the "new test" operator
([[ rather than [).