Generalize dependency test and keep track of path to MSG
Added a way to keep track of the path to MSG so that we don't have to
rely on a symlink in the working directory to the MSG directory.  The
only code change thereafter (i.e. in test_dependencies() of Utils.pm)
was to fix a bug revealed by calling msgCluster.pl directly, rather
than through a symlink in the working directory.

TODO: Check the rest of MSG's source for implicit usage of the working
directory symlink, and try to abstract it out.