Bug fix for structurally-varying parentals
Old code invalidly attempted to reference qname on dict rather than the
elements of the 'read' dict, which resulted in an uncaught exception in
certain instances, e.g. when a read maps forward to one parent, but
revcomp to the other, resulting in the seq field of the two SAMs
differing, but the read ID being the same.  Thanks to Elyssa Garza for
reporting this.