First attempt at adding support for BWA-MEM
Attempting to add support for BWA-MEM for both updating parentals and
mapping of cross individuals' reads.  Also added a lot of try-except
statements to attempt to handle newer versions of Pysam as well as
Pysam 0.1.2.  Handling deprecations up to at least 0.8.3 doesn't seem
mission-critical, as the old methods, attributes, and classes still
appear to work.  However, we don't know when backwards compatibility
will be dropped.

Noted during this review: Updating of parentals is likely the only
place where the consensus base is used in pileups.  This section of the
code is the limiting factor in allowing use of modern versions of
Samtools with MSG.