First attempt at adding support for BWA-MEM
BWA-MEM does not produce the same tags (e.g. XT, X0, X1) as BWA aln, so
when we add support, we need to use an alternate method of identifying
potentially repetitive mappings.  BWA-MEM provides the AS and XS tags,
and the difference between these (i.e. the difference between the score
of the primary alignment and the second-best alignment) lies in a
distribution where we can establish an empirical threshold for
repetitive mappings.  We choose AS-XS >= 6 here based on comparing the
AS-XS values for 1 Dsim individual's reads vs. those marked as XT:A:R
by bwa aln (see http://goo.gl/Ji771D ).  More empirical validation of
this threshold is pending though.