Filter hmmdata for multiple AIMs covered by one read
Added a feature to MSG that performs a different (less stringent)
heuristic filter on the hmmdata file to remove sites if they are within
a read length of the previous site.
This is an imperfect heuristic, but it is less stringent than the MSG
built-in, which filters all but one site in any collinear chain of
covered bases with gaps of <= 5 bp.

The ideal filter would be something like: reconstruct each read, and
filter any AIMs distal to the first AIM in the read.