Added configurable path to Samtools, plus permissions changes, and R package auto-installation
Added a variable in the configuration files for msgCluster.pl and
msgUpdateParentals.pl to allow for specification of a non-standard path
to the samtools executable, in case you want a version newer than 0.1.9
to be in your PATH, and still have MSG working.
Also adjusted permissions on the Bash, R, Python, and Perl scripts to
ensure executability by group and other, if needed.
Changed max_mapped_reads default value to 0 to make error messages more
informative.
Finally, added an R package automatic installation routine, rather than
naively checking presence.  This also validates loading of the packages.

Use 0 for max_mapped_reads rather than empty string

Changed the default to 0 (both in msgCluster.pl and msg.pl), and used a
numeric comparison with 0 for entering the SAM truncation code, rather
than checking for variable non-zero size.  On use of non-numeric value,
Bash should complain about the comparison being with non-numeric, which
is more informative (IMO).

Automatically install R packages and dependencies if needed

Rather than just testing for presence of packages, we can now check for
presence, loading, and version issues using require(), and attempt to
install the packages if they are not found.  If installation fails on
any one of them, we fail out of MSG before even starting by producing
an exit code of 1, which is passed to the calling Bash script, which is
again passed up all the way to msgCluster.pl.