Test all-in-one makefile
This makefile should compile the parts of MSG that need compiling,
including samtools-0.1.9, install the R packages, and provides a
directive for checking dependencies.
It doesn't actually add samtools to your PATH, in case you want a newer
version in your PATH.  But if you do that, be sure to specify the
samtools_path variable in your msg.cfg pointing to the samtools-0.1.9
installation that this makefile generates.