Solve CSV field error
Error "_csv.Error: field larger than field limit (131072)" is generated
when a field exceed 128 kB of text. While the exact origin of this error
is unknown, this could be due to high number of variants on a long
chromosome. The proposed solution sets the limit to the maximum possible
value.